requirements.txt
altair==4.1.0
appnope==0.1.0
argon2-cffi==20.1.0
astor==0.8.1
attrs==20.1.0
backcall==0.2.0
base58==2.0.1
bleach==3.1.5
blinker==1.4
blis==0.4.1
boto3==1.14.48
botocore==1.17.48
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fastai==2.0.0
fastcore==1.0.0
fastprogress==1.0.0
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.16.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.6.3
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.7
notebook==6.1.3
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
prometheus-client==0.8.0
prompt-toolkit==3.0.6
protobuf==3.13.0
ptyprocess==0.6.0
pyarrow==1.0.1
pycparser==2.20
pydeck==0.4.1
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.2
requests==2.24.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.2
Send2Trash==1.5.0
six==1.15.0
spacy==2.3.2
srsly==1.0.2
streamlit==0.65.2
terminado==0.8.3
testpath==0.4.4
thinc==7.4.1
threadpoolctl==2.1.0
toml==0.10.1
toolz==0.10.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.6.0+cpu
torchvision==0.7.0+cpu
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
tzlocal==2.1
urllib3==1.25.10
validators==0.18.0
wasabi==0.7.1
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.1.0
